+++ packages/date-picker/src/FilterDateRangePicker/types.spec.ts
@@ -1,6 +1,7 @@
 /* eslint-disable jest/expect-expect */
 import { DateRangeValidationStatus } from "./types"
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 let testValue: DateRangeValidationStatus
 
 describe("DateRangeValidationStatus", () => {